fix(utils): support values wrapped in a tuple
When a protobuf schema contains fields declared as a variant types (i.e.
oneof), the protobuf decoder returns a value wrapped in a tuple, where
the first element is a type name and the second one is the value. This
commit ensures that these tuple values can be encoded as json.

Fixes #9891