ci: add option to skip sanity checks
Sometimes, we just want to run the test cases or generate packages, but sanity checks are
breaking due to unrelated reasons.

This PR adds a label that skips those checks: `skip_sanity_checks`.

To avoid merging a PR with that label, we should make the `sanity_checks_passed` job a
required step for merging the PR.