feat: use preconditions to handle missing session data race condition
Specially when restarting a node, there might be a race condition when a client attempts
to connect and resume its session, but the shard its data lives in is not yet fully ready:
it might not yet have replicated/replayed all raft logs.  In such cases, the session will
appear as if it doesn't exist, and a fresh new one will be given the client.  After it
gets committed, we might get an inconsistent state where some parts of the state come from
the old session and other parts from the new one.  This problem stems from
`emqx_ds_replication_layer:get_streams`, which might return an empty stream list for a
shard that is not yet ready or has an RPC error.

To prevent this, we use preconditions to assert that there's indeed no session data when
creating the new one.  This will be executed by the raft state machine, so it'll be
consistent and does not rely on getting the latest streams list when attempting to open
the session.