fix: do not validate JWT iat
iat means "issued at", which is the unix epoch to record the time
when the token was issued.
previously, iat is not allowed to be in the future.
this cause some trouble for tokens issued in a host which
does not have clock synchronized.

given that the most important (security wise) claims are
nbf (not before) and exp (expire), iat validation has been removed.