fix: force keytab file to be the same as default value
The current sasl_auth implementation has a limit which only
allows serer side keytab file to be the same as system default
keytab file. Otherwise the authentication fails as soon as
server tries to validate the first auth packet from client.

This commit adds a config validator to ensure the configured
path is the same as current system default.