fix: ensure cluster link app can recover from init failure
Fixes EMQX-12868

If initial configuration for cluster link is incorrect, and local
cluster fails to establish a connection, an actor would crash on
startup which will lead to app supervisor crash, and removing of
cluster link config handler. This will make it impossible to remove or
update erroneous cluster link configuration.