fix(kafka consumer): deprecate `health_check_topic`
Fixes https://emqx.atlassian.net/browse/EMQX-13401

This field was accidentally added to the consumer schema because it was added to the
shared, v1 field in the shared schema module.  It never worked for the consumer connector,
as it uses a different library for interfacing with Kafka that does not need a topic to
health check.