fix(dsrepl): ensure preconditions are idemotent under log replays
When log replay is involved, preconditions should work as if they
observe the storage state at the specific point corresponding to
the Raft index, without the effects of any subsequent log entires
that may actually have been already applied to the storage state
before.

Because it's impossible with how we currently manage storage state
and snapshots, emulate that through keeping track of the latest Raft
index in the storage itself, and skip any (preconditions-involving)
batches that seem to be already applied by reading the latest Raft
index record on apply.