fix(wsconn): properly classify outgoing packets
Prior to this commit, outgoing packets ended up treated as "events",
making them postponed by appending to the process mailbox. As an
example, this caused the client to miss DISCONNECT packets, because
this behaviour was causing the WS socket to be closed _before_
DISCONNECT packet had a chance to be sent.