refactor(utils-json): drop legacy jsx-related (pre/post)processing
Motivated by the fact that nowadays in the overwhelming majority of cases
maps are used to represent JSON objects throughout the codebase. This
renders special preprocessing that turns deeply-nested-proplists into
`jiffy`-friendly representation wasteful. This commit drop this preprocessing,
and instead changes a few callsites that (often implicitly) rely on
proplists-friendliness.

Note that potential side-effects of this change loss of explicit JSON
field order.