refactor: check cluster join permission before join
pervisouly (merged to 5.9 but not released yet), the allow-join check
was done on the join requesting node, this may prevent node from
rolling upgrade without a license deployed first.

now the can_i_join rpc call will check if the requesting node is allowed
to join an existing node (with a valid license). The requesting node
will then copy the license after joined, so there is no need to add
license before jonning a licensed cluster.