perf(ws): avoid ets fullscan to keep track of max connections
Instead, use what amounts to a single `gen:call` to the connection
supervisor, which costs O(1) to process in the supervisor process
context. Under high load, this approach will still inevitably
compromise responsiveness of the supervisor. At least, this does
not waste enormous CPU resources, and can be thought of as
backpressure mechanism.