fix(ds): Fix beamformer queue handover procedure
This commit fixes high CPU load issue that may occur when a client is
subscribing to a topic in DS.

Fault scenario. Subscription is performed in three stages:

1. Catch-up to the tip of the stream in a specialized catch-up worker,
that is optimized for large batches

2. Hand-over to a "real-time" worker that is event-based and optimized
for low latency and data sharing

3. Then the subscription is kept in the real-time worker queue if the
client keeps up with the messages.

At step 2, the real-time worker verifies that the iterator is indeed
at the end of stream to avoid missing messages. If rate of messages
and new subscriptions is very high, this check can fail more and more
often, as hand-over is not instantaneous. This can create a
potentially infinite loop where the subscription is kicked back and
forth between the queues, which is costly.

This PR addresses this problem by changing the strategy of queue
handover in two ways:

1. If iterator received by RT queue  is lagging, RT worker will let it
catch up instead of refusing to serve it.

2. If catch up requires too much work, RT worker will drop the
subscription and signal the client instead of moving the iterator back
to the catchup queue.