feat(listener): support connection rate limiting for WS / WSS
Now both max connections limit and connection rate limit are enforced
through `ranch_conns_limiter` mechanism. This changes the behavior a
little: going over limit in either case now forces connections to be
closed much earlier, before HTTP and WebSocket protocol negotiation.
Which means the clients will now see it differently, as an abrupt
socket close. Clients are generally expected to be able to handle
such failures gracefully anyway, so it is not a breaking change.