perf(mt): make counter mnesia table and cache and monitor ets tables `ordered_set`
The original motivation for using `set` for `?COUNTER_TAB` was to be able to use
`mnesia:dirty_update_counter`.  But this function apparently works with `ordered_set` too:

https://github.com/emqx/otp/blob/7ebce9b1e8bb6b10637410600d1d65a5a9b79fc7/lib/mnesia/src/mnesia.erl#L2769

Since `ets:select` is efficient on `ordered_set` ETS tables when the key is partially
bound (such as fixing the namespace), we changed the types of those tables here for more
efficient scans/counts.

This shouldn't be a problem for rolling upgrades: the old table will not be created in the
new nodes even though it's part of the schema table, and the data in the old table does
not need to be migrated because the clients connected to the old node will simply
disconnect and eventually reconnect to new nodes, at which point they'll populate the new
counter table.