fix(ns config): use `sync_transaction` when saving configs
Attempt to fix https://emqx.atlassian.net/browse/EMQX-14728

Although I could not reproduce the reported issue (nor did the reporter), my hypothesis is
that something along the following lines occurred:

1) Connector was created on node A (using `mria:transaction`).
2) Action was created just after it.
3) When node B was replicating the action creation, it tried to check for the existence of
the connector, but mnesia had not yet replicated the config change to this node, and then
it crashed.

So, this is an attempt to reduce the chance of inconsistencies in a quick succession of
config changes.