fix(retainer): change retainer iteration flow to allow iteration when hitting rate limit
Fixes https://emqx.atlassian.net/browse/EMQX-14818

Fixes https://github.com/emqx/emqx/issues/15605

Ports the `client.handle_info` hook from 6.0 here.  Also, to make hook API more
extensible, changes the signature of the following hook callbacks to accept a more open
context map:

- `client.handle_info`
- `client.timeout`
- `message.delivered`
- `message.acked`

The context map, at this moment, has two keys which provide lazy access to channel and
session data: `chan_info_fn` and `session_info_fn`, which behave as `emqx_channel:info/2`
and `emqx_session:info/2` with the channel and session in the function's closure,
respectively.

This way, if we ever need to thread new data to one of these callbacks, we can simply add
it to the context map, and callbacks that are not interested in this new piece of
information do not need to be changed.